Code clean up.
Further code clean up.
Removed debug functions.
Fixed a bug with magazine reloading weapons - if reserve ammo was >= magazine capacity loaded, reload ended immediately upon loading ammo into the weapon instead of finishing the animation.

+++ src/main/java/com/mrcrayfish/guns/client/render/gun/model/MiniGunModel.java
@@ -59,7 +59,7 @@ public class MiniGunModel implements IOverrideModel
 		
 		if(rotationSpeed > 0F)
 		{
-			rotations.rotation += rotationSpeed;
+			rotations.rotation += (int) rotationSpeed;
 		}
 	}
