+++ src/main/java/com/mrcrayfish/guns/entity/grenade/ThrowableIncendiaryGrenadeEntity.java
@@ -56,7 +56,7 @@ public class ThrowableIncendiaryGrenadeEntity extends ThrowableGrenadeEntity
 		double speed = this.getDeltaMovement().length();
 		if(speed > 0.1)
 		{
-			this.rotation += (speed * 50);
+			this.rotation += (float) (speed * 50);
 		}
 	}
