+++ src/main/java/com/mrcrayfish/guns/entity/grenade/ThrowableMolotovEntity.java
@@ -73,7 +73,7 @@ public class ThrowableMolotovEntity extends ThrowableGrenadeEntity
 		double speed = this.getDeltaMovement().length();
 		if(speed > 0.1)
 		{
-			this.rotation += (speed * 50);
+			this.rotation += (float) (speed * 50);
 		}
 		if(this.isInWater())
 		{
